Added S3Controller to assist S3Agent
