+++ src/s3/S3Agent.ts
@@ -1,166 +1,16 @@
-import { fromEnv } from '@aws-sdk/credential-providers';
-import {
-  CopyObjectCommand,
-  DeleteObjectCommand,
-  GetObjectCommand,
-  HeadObjectCommand,
-  ListObjectsV2Command,
-  PutObjectCommand,
-  S3Client,
-} from '@aws-sdk/client-s3';
 import { Buffer } from 'buffer';
+import { S3Controller } from './S3Controller';
 
-// TODO-kev: ts complaining if set as variable: 'local' | 'staging' | 'prod'
-function getEndpoint(variable: string): string {
-  switch (variable) {
-    case 'local':
-      return 'http://localhost:9000/proavalon/';
-    case 'staging':
-      return 'https://s3-staging.proavalon.com/file/proavalon-staging/';
-    case 'prod':
-      return 'TO BE ADDED';
-  }
-}
-
-// TODO-kev: ts complaining if set as variable: 'local' | 'staging' | 'prod'
-function getBucket(variable: string): string {
-  switch (variable) {
-    case 'local':
-      return 'proavalon';
-    case 'staging':
-      return 'proavalon-staging';
-    case 'prod':
-      return 'TO BE ADDED';
-  }
+enum FolderName {
+  APPROVE = 'approved_avatars',
+  PENDING = 'pending_avatars',
 }
 
 class S3Agent {
-  private client: S3Client;
-  private endpoint: string;
-  private bucket: string;
+  private s3Controller: S3Controller;
 
   constructor() {
-    this.endpoint = getEndpoint(process.env.ENV);
-    this.bucket = getBucket(process.env.ENV);
-
-    if (process.env.ENV === 'local') {
-      this.client = new S3Client({
-        region: 'asdf',
-        endpoint: 'http://127.0.0.1:9000',
-        credentials: fromEnv(),
-      });
-    } else if (process.env.ENV == 'staging') {
-      this.client = new S3Client({
-        region: 'us-east-005',
-        endpoint: 'https://s3.us-east-005.backblazeb2.com',
-        credentials: fromEnv(),
-      });
-    }
-  }
-
-  public async objectExists(key: string) {
-    try {
-      const headCommand = new HeadObjectCommand({
-        Bucket: this.bucket,
-        Key: key,
-      });
-
-      await this.client.send(headCommand);
-      return true;
-    } catch (e) {
-      if (
-        e.name === 'NotFound' &&
-        e.$metadata &&
-        e.$metadata.httpStatusCode === 404
-      ) {
-        return false;
-      } else {
-        throw e;
-      }
-    }
-  }
-
-  private async listObjectKeys(prefixes: string[]) {
-    // Note ListObjects command only returns up to 1000 objects
-    // Need to update code if this exceeds
-    let keys: string[] = [];
-
-    for (const prefix of prefixes) {
-      const command = new ListObjectsV2Command({
-        Bucket: this.bucket,
-        Prefix: prefix,
-      });
-
-      const s3Objects = await this.client.send(command);
-
-      if (s3Objects.KeyCount !== 0) {
-        s3Objects.Contents.forEach((object) => {
-          keys.push(object.Key);
-        });
-      }
-    }
-
-    return keys;
-  }
-
-  public async uploadFile(key: string, fileContent: any, contentType: string) {
-    if (await this.objectExists(key)) {
-      throw new Error(`Failed to upload to s3. File already exists: '${key}'.`);
-    }
-
-    const command = new PutObjectCommand({
-      Bucket: this.bucket,
-      Key: key,
-      Body: fileContent,
-      ContentType: contentType,
-    });
-
-    await this.client.send(command);
-
-    console.log(`Successfully uploaded file to s3: ${key}`);
-  }
-
-  public async deleteObject(key: string) {
-    if (!(await this.objectExists(key))) {
-      console.log(
-        `Failed to delete. s3 file does not exist: ${this.bucket}/${key}.`,
-      );
-    }
-
-    const deleteCommand = new DeleteObjectCommand({
-      Bucket: this.bucket,
-      Key: key,
-    });
-
-    await this.client.send(deleteCommand);
-    console.log(`Successfully deleted s3 file: ${this.bucket}/${key}.`);
-  }
-
-  public async refactorObjectFilepath(oldKey: string, newKey: string) {
-    if (!(await this.objectExists(oldKey))) {
-      throw new Error(
-        `Failed to refactor s3 file. Object with key '${oldKey}' does not exist.`,
-      );
-    }
-
-    if (await this.objectExists(newKey)) {
-      throw new Error(
-        `Failed to refactor s3 file. Destination object with key '${newKey}' already exists.`,
-      );
-    }
-
-    const copyCommand = new CopyObjectCommand({
-      Bucket: this.bucket,
-      CopySource: `${this.bucket}/${oldKey}`,
-      Key: newKey,
-    });
-
-    await this.client.send(copyCommand);
-    await this.deleteObject(oldKey);
-
-    console.log(
-      `Successfully refactored filepath in s3 from: ${this.bucket}/${oldKey} to: ${this.bucket}/${newKey}`,
-    );
+    this.s3Controller = new S3Controller(process.env.ENV);
   }
 
   // =====================================================
@@ -170,18 +20,20 @@ class S3Agent {
   // Uploads avatar requests to s3. Presumes validation checks have been completed
   // Returns accessible links for res and spy avatars
   // Note to add a new prefix for all files to be checked
-  // TODO-kev check File type
   public async uploadAvatarRequestImages(
     username: string,
     resAvatar: Buffer,
     spyAvatar: Buffer,
   ) {
     const usernameLower = username.toLowerCase();
-    const prefix1 = `pending_avatars/${usernameLower}/`;
-    const prefix2 = `approved_avatars/${usernameLower}/`;
+    const prefix1 = `${FolderName.PENDING}/${usernameLower}/`;
+    const prefix2 = `${FolderName.APPROVE}/${usernameLower}/`;
 
     // Find unique ID for filepath generation
-    const existingObjectKeys = await this.listObjectKeys([prefix1, prefix2]);
+    const existingObjectKeys = await this.s3Controller.listObjectKeys([
+      prefix1,
+      prefix2,
+    ]);
     const counter = this.getCurrentKeyCounter(existingObjectKeys);
 
     const resFileName = `${usernameLower}_res_${counter + 1}.png`;
@@ -193,21 +45,21 @@ class S3Agent {
     const succeededFiles = [];
 
     try {
-      await this.uploadFile(resKey, resAvatar, 'image/png');
+      await this.s3Controller.uploadFile(resKey, resAvatar, 'image/png');
       succeededFiles.push(resKey);
 
-      await this.uploadFile(spyKey, spyAvatar, 'image/png');
+      await this.s3Controller.uploadFile(spyKey, spyAvatar, 'image/png');
       succeededFiles.push(spyKey);
     } catch (e) {
       for (const succeededFile of succeededFiles) {
-        await this.deleteObject(succeededFile);
+        await this.s3Controller.deleteObject(succeededFile);
       }
 
       throw e;
     }
 
-    const resLink = `${this.endpoint}${resKey}`;
-    const spyLink = `${this.endpoint}${spyKey}`;
+    const resLink = `${this.s3Controller.getEndpoint()}${resKey}`;
+    const spyLink = `${this.s3Controller.getEndpoint()}${spyKey}`;
 
     return { resLink, spyLink };
   }
@@ -221,9 +73,10 @@ class S3Agent {
 
       // TODO-kev: Thoughts on this one?
       // Match format <pending_avatars|approved_avatars>/<username>/<username>_<res|spy>_<id>.png
-      const match = key.match(
-        /^(pending_avatars|approved_avatars)\/([^/]+)\/\2_(res|spy)_(\d+)\.png$/,
+      const pattern = new RegExp(
+        `^(${FolderName.PENDING}|${FolderName.APPROVE})\\/([^/]+)\\/\\2_(res|spy)_(\\d+)\\.png$`,
       );
+      const match = key.match(pattern);
 
       if (match) {
         const count = parseInt(match[4], 10);
@@ -237,24 +90,28 @@ class S3Agent {
   }
 
   public async rejectAvatarRequest(link: string) {
-    const pattern = /pending_avatars\/.*$/;
+    const pattern = new RegExp(`${FolderName.PENDING}\/.*$`);
     if (!link.match(pattern)) {
       throw new Error(`Invalid link provided: ${link}`);
     }
 
     const key = link.match(pattern)[0];
+    console.log(key);
 
-    await this.deleteObject(key);
+    await this.s3Controller.deleteObject(key);
   }
 
   public async approveAvatarRefactorFilePath(link: string) {
     // TODO-kev: Consider usernames pending_avatars
-    const index = link.indexOf('pending_avatars');
+    const index = link.indexOf(FolderName.PENDING);
     const endpoint = link.substring(0, index);
     const key = link.substring(index);
-    const newKey = key.replace('pending_avatars', 'approved_avatars');
+    const newKey = key.replace(
+      `${FolderName.PENDING}`,
+      `${FolderName.APPROVE}`,
+    );
 
-    await this.refactorObjectFilepath(key, newKey);
+    await this.s3Controller.refactorObjectFilepath(key, newKey);
     return `${endpoint}${newKey}`;
   }
 }
